feat: api, modal ingredient
