+++ src/components/BurgerIngredients/BurgerIngredient/BurgerIngredient.jsx
@@ -3,23 +3,17 @@ import {
   Counter,
 } from "@ya.praktikum/react-developer-burger-ui-components";
 import burgerStyle from "./BurgerIngredient.module.css";
-import { useState, useCallback } from "react";
-import Modal from "../../Modal/Modal";
-import IngredientDetails from "./IngredientDetails/IngredientDetails";
-import { useSelector } from "react-redux";
+import { useCallback } from "react";
 import { useDrag } from "react-dnd/dist/hooks";
 import {
   GET_INGREDIENT_INFO,
-  DELETE_INGREDIENT_INFO,
 } from "../../../services/actions/ingredientInfo";
 import { useDispatch } from "react-redux";
 import PropTypes from "prop-types";
 import { ingredientType } from "../../../utils/types";
 
-const BurgerIngredient = ({ count, item }) => {
+const BurgerIngredient = ({ count, item, open }) => {
   const dispatch = useDispatch();
-  const [modalActive, setModalActive] = useState(false);
-  const ingredient = useSelector((state) => state.ingredientInfo.item);
   const [{ opacity }, dragRef] = useDrag({
     type: "ingredients",
     item,
@@ -30,14 +24,8 @@ const BurgerIngredient = ({ count, item }) => {
 
   const onClickImage = useCallback(() => {
     dispatch({ type: GET_INGREDIENT_INFO, item });
-    setModalActive(true);
+    open()
   }, [dispatch, item]);
-
-  // Функция закрытия модального окна
-  function onClose() {
-    setModalActive(false);
-    dispatch({ type: DELETE_INGREDIENT_INFO });
-  }
 
   return (
     <>
@@ -60,11 +48,6 @@ const BurgerIngredient = ({ count, item }) => {
         </div>
         <p className="text text_type_main-default mt-2">{item.name}</p>
       </div>
-      {modalActive && ingredient && (
-        <Modal title={"Детали ингредиента"} onClose={onClose}>
-          <IngredientDetails />
-        </Modal>
-      )}
     </>
   );
 };
