+++ src/utils/api.js
@@ -5,22 +5,25 @@ const apiConfig = {
   },
 };
 
+function request(url, options) {
+  return fetch(url, options).then(checkResponse);
+}
+
 const checkResponse = (res) => {
   return res.ok ? res.json() : Promise.reject(res);
 };
 
-export const getData =  () => {
-  return fetch(`${apiConfig.baseURL}ingredients`,{
-  method: "GET",
-  headers: apiConfig.headers,
-}).then(checkResponse);
+export const getData = () => {
+  return request(`${apiConfig.baseURL}ingredients`, {
+    method: "GET",
+    headers: apiConfig.headers,
+  });
 };
 
-export const apiOrder =  (orderInfo) => {
- return fetch(`${apiConfig.baseURL}orders`, {
+export const apiOrder = (orderInfo) => {
+  return request(`${apiConfig.baseURL}orders`, {
     method: "POST",
     headers: apiConfig.headers,
     body: JSON.stringify({ ingredients: orderInfo }),
-  })
-  .then(checkResponse);
+  });
 };
